delete success

+++ src/main/java/tn/iit/teacher_authorization_management/dao/ProfessorDAO.java
@@ -75,7 +75,7 @@ public class ProfessorDAO {
 		}
 	}
 
-	public void deleteProfessor(int professorId) {
+	public void deleteProfessor(Long professorId) {
 	    Session session = sessionFactory.openSession();
 	    Transaction transaction = null;
 	    try {
